+++ gdtoolkit/formatter/annotation.py
@@ -1,10 +1,11 @@
 from lark import Tree
 
-from .types import FormattedLine, FormattedLines
-from .context import Context
+from .types import FormattedLine, FormattedLines, Outcome
+from .context import Context, ExpressionContext
+from .expression import format_concrete_expression
 from .expression_to_str import expression_to_str
 
-STANDALONE_ANNOTATIONS = ["tool"]
+STANDALONE_ANNOTATIONS = ["tool", "icon"]
 
 
 def is_non_standalone_annotation(statement: Tree) -> bool:
@@ -33,16 +34,16 @@ def prepend_annotations_to_formatted_line(
         return [(line_to_prepend_to[0], single_line)]
     formatted_lines: FormattedLines = []
     for annotation in context.annotations:
-        formatted_lines.append(format_annotation_to_line(annotation, context))
+        lines, _ = format_standalone_annotation(annotation, context)
+        formatted_lines += lines
     formatted_lines.append(line_to_prepend_to)
     context.annotations = []
     return formatted_lines
 
 
-def format_annotation_to_line(annotation: Tree, context: Context) -> FormattedLine:
-    return (
-        annotation.line,
-        "{}{}".format(context.indent_string, format_annotation_to_string(annotation)),
+def format_standalone_annotation(annotation: Tree, context: Context) -> Outcome:
+    return format_concrete_expression(
+        annotation, ExpressionContext("", -1, "", -1), context
     )